+++ craft_application/commands/lifecycle.py
@@ -405,11 +405,19 @@ class PackCommand(LifecycleCommand):
 
         if not packages:
             emit.progress("No packages created.", permanent=True)
+            artifact, resources = None, None
         elif len(packages) == 1:
             emit.progress(f"Packed {packages[0].name}", permanent=True)
+            artifact, resources = packages[0], None
         else:
             package_names = ", ".join(pkg.name for pkg in packages)
             emit.progress(f"Packed: {package_names}", permanent=True)
+            artifact, resources = packages[0], self._services.package.resource_map
+
+        # This will use the provider shared registry after it's implemented
+        # in craft-providers, to allow transparent data transfer from instances
+        # or unmanaged runs.
+        self._services.package.write_state(artifact=artifact, resources=resources)
 
         if shell_after:
             _launch_shell()
@@ -474,8 +482,19 @@ class TestCommand(PackCommand):
             # Run the rest of this outside the managed instance.
             return
 
+        # This will use the provider shared registry after it's implemented
+        # in craft-providers, to allow transparent data transfer from instances
+        # or unmanaged runs.
+        if parsed_args.destructive_mode:
+            pack_state = self._services.package.read_state()
+        else:
+            pack_state = self._services.provider.get_pack_state()
+
+        if not pack_state.artifact:
+            raise RuntimeError("No artifact files to test.")
+
         emit.progress("Testing project")
-        self._services.get("testing").test(pathlib.Path.cwd())
+        self._services.get("testing").test(pathlib.Path.cwd(), pack_state)
         emit.progress("Tests succeeded")